Testing out touch events in the GLView
